No longer a need for a separate addTextArea in the frontend form class
